docs: change copyright holder in `LICENSE` and copyright notices (#82)
Change the `LICENSE` and copyright notices to reflect the current copyright holder.

BREAKING CHANGE: Copyright holder changed

The copyright holder changed. If you have a copy of `LICENSE` in your local copy of the software, please update it to reflect the new copyright holder, when updating to the new version.

* ci: use fork of `sync-branch`

Use a fork of action `sync-branch`.

* docs: change copyright holder in `LICENSE` and copyright notices

Change the `LICENSE` and copyright notices to reflect the current copyright holder.